+++ tests/end2end_test/g_spreadsheet_end_to_end_test.py
@@ -10,7 +10,7 @@ from dags.google_spreadsheet_import_pipeline import (
     DEFAULT_DEPLOYMENT_ENV_VALUE, DEPLOYMENT_ENV_ENV_NAME
 )
 from data_pipeline.utils.pipeline_file_io import get_yaml_file_as_dict
-from data_pipeline.spreadsheet_data.google_spreadsheet_config import (
+from data_pipeline.google_spreadsheet.google_spreadsheet_config import (
     MultiSpreadsheetConfig, MultiCsvSheet
 )
 from tests.end2end_test import enable_and_trigger_dag_and_wait_for_success
